modules/cloud_function/source/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-build==3.*
PyYAML==6.*
